+++ src/attribute.ts
@@ -127,11 +127,11 @@ export abstract class AttributeToken extends Token {
 	}
 
 	/** attribute value / 属性值 */
-	get value(): string | true {
+	get value(): string {
 		return this.getValue();
 	}
 
-	set value(value) {
+	set value(value: string | true) {
 		this.setValue(value);
 	}
 
@@ -308,7 +308,7 @@ export abstract class AttributeToken extends Token {
 					return e;
 				}
 			}
-		} else if (name === 'style' && typeof value === 'string' && insecureStyle.test(value)) {
+		} else if (name === 'style' && insecureStyle.test(value)) {
 			/* PRINT ONLY */
 
 			if (start === undefined) {
@@ -322,7 +322,7 @@ export abstract class AttributeToken extends Token {
 				const s = lintConfig.getSeverity(rule);
 				return s && generateForChild(lastChild, rect!, rule, 'insecure-style', s);
 			}
-		} else if (name === 'tabindex' && typeof value === 'string' && value !== '0') {
+		} else if (name === 'tabindex' && value !== '0') {
 			/* PRINT ONLY */
 
 			if (start === undefined) {
@@ -345,13 +345,11 @@ export abstract class AttributeToken extends Token {
 				}
 			}
 		} else if (
-			typeof value === 'string' && (
-				(/^xmlns:[\w:.-]+$/u.test(name) || urlAttrs.has(name)) && evil.test(value)
-				|| simple
-				&& (name === 'href' || type === 'ext-attr' && tag === 'img' && name === 'src')
-				&& !new RegExp(String.raw`^(?:${this.getAttribute('config').protocol}|//)\S+$`, 'iu')
-					.test(value)
-			)
+			(/^xmlns:[\w:.-]+$/u.test(name) || urlAttrs.has(name)) && evil.test(value)
+			|| simple
+			&& (name === 'href' || type === 'ext-attr' && tag === 'img' && name === 'src')
+			&& !new RegExp(String.raw`^(?:${this.getAttribute('config').protocol}|//)\S+$`, 'iu')
+				.test(value)
 		) {
 			/* PRINT ONLY */
 
@@ -367,7 +365,7 @@ export abstract class AttributeToken extends Token {
 			}
 		} else if (simple && type !== 'ext-attr') {
 			const data = provideValues(tag, name),
-				v = String(value).toLowerCase();
+				v = value.toLowerCase();
 			if (data.length > 0 && data.every(n => n !== v)) {
 				/* PRINT ONLY */
 
@@ -460,8 +458,8 @@ export abstract class AttributeToken extends Token {
 	 *
 	 * 获取属性值
 	 */
-	getValue(): string | true {
-		return this.#equal ? this.lastChild.text().trim() : this.type === 'ext-attr' || '';
+	getValue(): string {
+		return this.#equal ? this.lastChild.text().replace(/[\t\r\n ]+/gu, ' ').trim() : '';
 	}
 
 	/** @private */