+++ src/attributes.ts
@@ -118,19 +118,18 @@ export abstract class AttributesToken extends Token {
 	/* NOT FOR BROWSER */
 
 	/** all attributes / 全部属性 */
-	get attributes(): Record<string, string | true> {
+	get attributes(): Record<string, string> {
 		return this.getAttrs();
 	}
 
-	set attributes(attrs) {
+	set attributes(attrs: Record<string, string | true>) {
 		this.replaceChildren();
 		this.setAttr(attrs);
 	}
 
 	/** class attribute in string / 以字符串表示的class属性 */
 	get className(): string {
-		const attr = this.getAttr('class');
-		return typeof attr === 'string' ? attr : '';
+		return this.getAttr('class') || '';
 	}
 
 	set className(className) {
@@ -164,8 +163,7 @@ export abstract class AttributesToken extends Token {
 
 	/** id attribute / id属性 */
 	get id(): string {
-		const attr = this.getAttr('id');
-		return typeof attr === 'string' ? attr : '';
+		return this.getAttr('id') || '';
 	}
 
 	set id(id) {
@@ -299,7 +297,7 @@ export abstract class AttributesToken extends Token {
 	 * 获取指定属性
 	 * @param key attribute name / 属性键
 	 */
-	getAttr(key: string): string | true | undefined {
+	getAttr(key: string): string | undefined {
 		LINT: return this.getAttrToken(key)?.getValue();
 	}
 
@@ -358,12 +356,8 @@ export abstract class AttributesToken extends Token {
 				const severity = lintConfig.getSeverity(rules[1], tag);
 				if (severity) {
 					for (const key of required.get(tag)!) {
-						const keys = typeof key === 'string' ? [key] : key,
-							missing = keys.every(k => {
-								const value = this.getAttr(k);
-								return value === true || !value;
-							});
-						if (missing) {
+						const keys = typeof key === 'string' ? [key] : key;
+						if (!keys.some(k => this.getAttr(k))) {
 							errors.push(generateForSelf(
 								this,
 								rect,
@@ -378,10 +372,7 @@ export abstract class AttributesToken extends Token {
 			const severity = lintConfig.getSeverity(rules[2], 'attribute');
 			if (severity && duplicated.size > 0) {
 				for (const key of duplicated) {
-					const pairs = attrs.get(key)!.map(attr => {
-						const value = attr.getValue();
-						return [attr, value === true ? '' : value] as const;
-					});
+					const pairs = attrs.get(key)!.map(attr => [attr, attr.getValue()] as const);
 					Array.prototype.push.apply(
 						errors,
 						pairs.map(([attr, value], i) => {
@@ -538,7 +529,7 @@ export abstract class AttributesToken extends Token {
 	 *
 	 * 获取全部属性
 	 */
-	getAttrs(): Record<string, string | true> {
+	getAttrs(): Record<string, string> {
 		return Object.fromEntries(this.getAttrTokens().map(({name, value}) => [name, value]));
 	}