+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Lowered severity of a lonely `]` following a free external link and some plain text
 - Update PMID magic link URLs
+- [`AttributeToken#getValue`](https://github.com/bhsd-harry/wikiparser-node/wiki/AttributeToken-%28EN%29#getvalue) should always return a string and normalize whitespace characters to a single space
 
 ## v1.41.0